\name{ppiAIPW-package}
\alias{ppiAIPW-package}
\alias{ppiAIPW}
\title{Semi-Supervised Mean Inference with AIPW and Calibration}
\description{
Native R implementation of semi-supervised mean inference with AIPW-style
augmentation, calibration methods, automatic method selection, diagnostics,
and a causal convenience wrapper.
}
\docType{package}
\author{
Lars van der Laan
}
\keyword{package}
